redefine interactions
