+++ docs/api/compas_model.interactions.rst
@@ -11,4 +11,5 @@ Classes
     :toctree: generated/
     :nosignatures:
 
-    Interaction
+    Contact
+    Modifier
